pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "invenio-subjects-cessda"
dynamic = ["version"]
description = "CESSDA subject vocabularies for InvenioRDM."
readme = {file = "README.md", content-type = "text/markdown"}
keywords = ["invenio", "vocabularies", "cessda"]
license = {text = "MIT"}
authors = [
    {name = "KTH Royal Institute of Technology", email = "info@kth.se"},
]
requires-python = ">=3.9"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = []

[project.optional-dependencies]
tests = [
    "pytest>=7.0.0",
    "pytest-invenio>=2.1.7",
    "ruff>=0.2.0",
    "aiohttp>=3.9.0",
    "PyYAML>=6.0.1",
    "click>=8.1.7",
]

[project.entry-points."invenio_rdm_records.fixtures"]
invenio_subjects_cessda = "invenio_subjects_cessda.vocabularies"

[project.urls]
Homepage = "https://github.com/Samk13/invenio-subjects-cessda"
Repository = "https://github.com/Samk13/invenio-subjects-cessda"
Issues = "https://github.com/Samk13/invenio-subjects-cessda/issues"

[tool.hatch.version]
path = "invenio_subjects_cessda/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["invenio_subjects_cessda"]
exclude = [
    "main.py",
    "run-tests.sh",
    "Makefile",
    "tests",
    "README.md",
    "MANIFEST.in",
    ".github",
    ".vscode",
    "uv.lock",
    ".pylintrc",
    ".gitignore",
    "invenio_subjects_cessda/config.py",
    "invenio_subjects_cessda/convert.py",
    "invenio_subjects_cessda/delta.py",
    "invenio_subjects_cessda/fetch_voc.py",
    "invenio_subjects_cessda/schemas.py",
    "invenio_subjects_cessda/utils.py",
    "invenio_subjects_cessda/writer.py",
]

[tool.hatch.build.targets.sdist]
exclude = [
    "main.py",
    "run-tests.sh",
    "Makefile",
    "tests",
    "README.md",
    "MANIFEST.in",
    ".github",
    ".vscode",
    "uv.lock",
    ".pylintrc",
    ".gitignore",
    "invenio_subjects_cessda/config.py",
    "invenio_subjects_cessda/convert.py",
    "invenio_subjects_cessda/delta.py",
    "invenio_subjects_cessda/fetch_voc.py",
    "invenio_subjects_cessda/schemas.py",
    "invenio_subjects_cessda/utils.py",
    "invenio_subjects_cessda/writer.py",
]

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "SIM",
]
ignore = [
    "E501",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["invenio_subjects_cessda"]
combine-as-imports = true

[tool.pytest.ini_options]
addopts = "--doctest-glob='*.rst' --doctest-modules"
filterwarnings = ["ignore::pytest.PytestDeprecationWarning"]
testpaths = ["tests", "invenio_subjects_cessda"]
